+++ changelog.md
@@ -2,10 +2,20 @@
 
 **Test Status**
 * G - Controller Files
-	* 'Connected Device Class' is recognized successfully.
-	* Actual function tests still fail.
+	* Started writing emulation for 'Connected Device Class' tests.
+	* So far, only object type validation is successful.
+	* Property type validation fails.
+
+---
+
+**./fox-devices/_classes/device-model.class.js**
+* Added notes referring to class SET methods.
 
 ---
 
 **./fox-devices/_classes/device.class.js**
-* Wrote basic class structure.
+* Added requirement for '../../fox-custom/validation-tasks'
+* If the input object is not a 'StoredDevice', an error will be displayed.
+* Declared new function 'handlePropertyValidation'
+	* This will handle property validation.
+	* So far, only 'id' is checked.